+++ changelog.md
@@ -5,6 +5,7 @@
 #### New features
 
 * Added function `Kymo.line_timestamp_ranges()` to obtain the start and stop timestamp of each scan line in a `Kymo`. Please refer to [Confocal images](https://lumicks-pylake.readthedocs.io/en/latest/tutorial/kymographs.html) for more information.
+* Added `Kymo.flip()` to flip a Kymograph along its positional axis.
 
 #### Breaking changes